requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.3.1
Django==1.11.15
django-ranged-response==0.2.0
django-simple-captcha==0.5.9
idna==2.7
Pillow==5.2.0
pycparser==2.19
PyMySQL==0.9.2
pytz==2018.5
six==1.11.0
